+++ HISTORY.rst
@@ -8,6 +8,9 @@ Release History
 
 - Added message send retry.
 - Added timeouts and better error handling for management requests.
+- Improved connection and auth error handling.
+- Fixed message annotations type.
+- SendClient.send_all_messages() now returns a list of message send statuses.
 
 
 0.1.0b1 (2018-03-24)